+++ tests/test_model.py
@@ -1,15 +1,15 @@
 import pytest
 
-from supertask.model import CronJob
+from supertask.model import ScheduleItem
 
 
-def test_cronjob_fail_5():
+def test_cronjob_fail_4():
     """
     Validate decoding fails because crontab syntax includes too few fields.
     """
-    job = CronJob(id=42, name="foobar", trigger_cron="1 2 3 4")
+    schedule = ScheduleItem(cron="1 2 3 4")
     with pytest.raises(ValueError) as ex:
-        job.decode_crontab()
+        schedule.crontab()
     assert ex.match("Invalid crontab syntax: 1 2 3 4")
 
 
@@ -17,9 +17,9 @@ def test_cronjob_fail_8():
     """
     Validate decoding fails because crontab syntax includes too many fields.
     """
-    job = CronJob(id=42, name="foobar", trigger_cron="1 2 3 4 5 6 7 8")
+    schedule = ScheduleItem(cron="1 2 3 4 5 6 7 8")
     with pytest.raises(ValueError) as ex:
-        job.decode_crontab()
+        schedule.crontab()
     assert ex.match("Invalid crontab syntax: 1 2 3 4 5 6 7 8")
 
 
@@ -27,8 +27,8 @@ def test_cronjob_success_5():
     """
     Validate decoding succeeds with 5 fields worth of crontab syntax.
     """
-    job = CronJob(id=42, name="foobar", trigger_cron="* * * * *")
-    crontab = job.decode_crontab()
+    schedule = ScheduleItem(cron="* * * * *")
+    crontab = schedule.crontab()
     assert crontab["second"] is None
     assert crontab["minute"] == "*"
     assert crontab["year"] is None
@@ -38,8 +38,8 @@ def test_cronjob_success_6():
     """
     Validate decoding succeeds with 5 fields worth of crontab syntax, including seconds.
     """
-    job = CronJob(id=42, name="foobar", trigger_cron="*/10 * * * * *")
-    crontab = job.decode_crontab()
+    schedule = ScheduleItem(cron="*/10 * * * * *")
+    crontab = schedule.crontab()
     assert crontab["second"] == "*/10"
     assert crontab["minute"] == "*"
     assert crontab["year"] is None
@@ -49,8 +49,8 @@ def test_cronjob_success_7():
     """
     Validate decoding succeeds with 5 fields worth of crontab syntax, including year.
     """
-    job = CronJob(id=42, name="foobar", trigger_cron="*/10 * * * * * 2026")
-    crontab = job.decode_crontab()
+    schedule = ScheduleItem(cron="*/10 * * * * * 2026")
+    crontab = schedule.crontab()
     assert crontab["second"] == "*/10"
     assert crontab["minute"] == "*"
     assert crontab["year"] == "2026"